tests/wpt/web-platform-tests/css/filter-effects/parsing/lighting-color-parsing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: parsing lighting-color with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#LightingColorProperty">
<meta name="assert" content="lighting-color supports only the grammar '<color>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("lighting-color", "auto");
test_invalid_value("lighting-color", "none");
test_invalid_value("lighting-color", "#a");
</script>
</body>
</html>
